use cents in internal models

+++ src/main/kotlin/kpavlov/bank/api/model/AccountStatement.kt
@@ -2,13 +2,12 @@ package kpavlov.bank.api.model
 
 import kpavlov.bank.domain.AccountId
 import kpavlov.bank.domain.AccountType
-import java.math.BigDecimal
 import java.time.OffsetDateTime
 
 data class AccountStatement(
         val id: AccountId,
         val type: AccountType,
-        val balance: BigDecimal,
+        val balance: Long,
         val transactions: List<Transaction>,
         val timestamp: OffsetDateTime
 )
