+++ src/main/kotlin/kpavlov/bank/api/model/CustomerDetails.kt
@@ -1,12 +1,11 @@
 package kpavlov.bank.api.model
 
 import kpavlov.bank.domain.CustomerId
-import java.math.BigDecimal
 
 data class CustomerDetails(
         val id: CustomerId,
         val firstName: String,
         val lastName: String,
-        val balance: BigDecimal,
+        val balance: Long,
         val accounts: List<AccountStatement>
 )
